requirements.txt
baukit==0.0.1
numpy==1.24.3
pandas==2.0.2
peft==0.10.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
torch==1.13.1
tqdm==4.65.0
transformers==4.34.0
